initial support for undo/redo of inline constants
currently this is not very nice, because there is no external way to tell the pin widgets to update their inline values from the graph. The only way to make it happen is to force a full rebuild of the node. InlinePinWidget.UpdateInputFromModifiedTextEntry() needs to be refactored, and possibly there needs to be a way for the graph/node to signal constant-value changes

(also deeper issue w/ constant-value vs default value...)